Support for long primary index on 32 bit devices

+++ RBQFetchedResultsController/CacheObjects/RBQControllerCacheObject.m
@@ -11,7 +11,7 @@
 @implementation RBQControllerCacheObject
 
 + (instancetype)cacheWithName:(NSString *)name
-             fetchRequestHash:(NSInteger)hash
+             fetchRequestHash:(long long)hash
 {
     RBQControllerCacheObject *cache = [[RBQControllerCacheObject alloc] init];
     cache.name = name;
